requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
requests
pyotp
qrcode
pillow
